pyproject.toml
[project]
name = "cait"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "litellm>=1.57.7",
    "paramiko>=3.5.0",
    "python-dotenv>=1.0.1",
    "smolagents>=1.2.2",
    "watchdog>=6.0.0",
]
